fix: address PR feedback items
