+++ app/assets/javascripts/components/Button.tsx
@@ -1,10 +1,10 @@
 import { FunctionComponent } from 'preact';
 
-const base = `rounded px-4 py-1.75 font-bold text-sm fit-content cursor-pointer`;
+const baseClass = `rounded px-4 py-1.75 font-bold text-sm fit-content cursor-pointer`;
 
-const normal = `${base} bg-default color-text border-solid border-gray-300 border-1 \
+const normalClass = `${baseClass} bg-default color-text border-solid border-gray-300 border-1 \
 focus:bg-contrast hover:bg-contrast`;
-const primary = `${base} no-border bg-info color-info-contrast hover:brightness-130 \
+const primaryClass = `${baseClass} no-border bg-info color-info-contrast hover:brightness-130 \
 focus:brightness-130`;
 
 export const Button: FunctionComponent<{
@@ -13,7 +13,7 @@ export const Button: FunctionComponent<{
   label: string;
   onClick: () => void;
 }> = ({ type, label, className = '', onClick }) => {
-  const buttonClass = type === 'primary' ? primary : normal;
+  const buttonClass = type === 'primary' ? primaryClass : normalClass;
   return (
     <button
       className={`${buttonClass} ${className}`}
