+++ app/assets/javascripts/preferences/PreferencesView.tsx
@@ -3,8 +3,8 @@ import { TitleBar, Title } from '@/components/TitleBar';
 import { FunctionComponent } from 'preact';
 import { HelpAndFeedback, Security } from './panes';
 import { observer } from 'mobx-react-lite';
-import { MenuItem } from './components';
 import { PreferencesMenu } from './preferences-menu';
+import { PreferencesMenuView } from './PreferencesMenuView';
 
 const PaneSelector: FunctionComponent<{
   prefs: PreferencesMenu;
@@ -40,19 +40,12 @@ const PreferencesCanvas: FunctionComponent<{
   </div>
 ));
 
-const blockPropagation = (e: MouseEvent) => {
-  e.stopPropagation();
-};
-
 const PreferencesView: FunctionComponent<{ close: () => void }> = observer(
   ({ close }) => {
     const prefs = new PreferencesMenu();
 
     return (
-      <div
-        onClick={blockPropagation}
-        className="sn-full-screen flex flex-col bg-contrast z-index-preferences"
-      >
+      <div className="sn-full-screen flex flex-col bg-contrast z-index-preferences">
         <TitleBar className="items-center justify-between">
           {/* div is added so flex justify-between can center the title */}
           <div className="h-8 w-8" />
@@ -82,19 +75,3 @@ export const PreferencesViewWrapper: FunctionComponent<PreferencesWrapperProps> 
       <PreferencesView close={() => appState.preferences.closePreferences()} />
     );
   });
-
-export const PreferencesMenuView: FunctionComponent<{
-  menu: PreferencesMenu;
-}> = observer(({ menu }) => (
-  <div className="min-w-55 overflow-y-auto flex flex-col px-3 py-6">
-    {menu.menuItems.map((pref) => (
-      <MenuItem
-        key={pref.id}
-        iconType={pref.icon}
-        label={pref.label}
-        selected={pref.selected}
-        onClick={() => menu.selectPane(pref.id)}
-      />
-    ))}
-  </div>
-));
