+++ app/assets/javascripts/preferences/panes/two-factor-auth/TwoFactorDialog.tsx
@@ -18,8 +18,15 @@ export const TwoFactorDialog: FunctionComponent<{
 
   return (
     <AlertDialog leastDestructiveRef={ldRef}>
-      <div className="sn-component">
-        <div className="w-160 bg-default rounded shadow-overlay focus:padded-ring-info">
+      {/* sn-component is focusable by default, but doesn't stretch to child width
+          resulting in a badly focused dialog. Utility classes are not available 
+          at the sn-component level, only below it. tabIndex -1 disables focus 
+          and enables it on the child component */}
+      <div tabIndex={-1} className="sn-component">
+        <div
+          tabIndex={0}
+          className="w-160 bg-default rounded shadow-overlay focus:padded-ring-info"
+        >
           {children}
         </div>
       </div>
